requirements.txt
pyserial
ollama
langgraph
langchain-community
langchain-core